refactor: Cost Breakdown component
