+++ public/static/locales/en-US/checkout.json
@@ -14,18 +14,7 @@
     }
   },
   "costBreakdown": {
-    "loading": "Loading your cart…",
-    "thead": {
-      "amount": "Qty",
-      "item": "Item",
-      "price": "Price"
-    },
-    "tbody": {
-      "shipping": "Shipping cost"
-    },
-    "tfoot": {
-      "total": "Total"
-    }
+    "loading": "Loading your cart…"
   },
   "contact": {
     "title": "Contact for delivery",
