+++ public/static/locales/th/checkout.json
@@ -14,18 +14,7 @@
     }
   },
   "costBreakdown": {
-    "loading": "กำลังโหลดรถเข็นของคุณ…",
-    "thead": {
-      "amount": "จำนวน",
-      "item": "สินค้า",
-      "price": "ราคา"
-    },
-    "tbody": {
-      "shipping": "ค่าขนส่ง"
-    },
-    "tfoot": {
-      "total": "รวม"
-    }
+    "loading": "กำลังโหลดรถเข็นของคุณ…"
   },
   "contact": {
     "title": "ช่องทางการติดต่อสำหรับขนส่ง",
