+++ public/static/locales/th/receipt.json
@@ -1,0 +1,16 @@
+{
+  "costBreakdown": {
+    "loading": "กำลังโหลดรถเข็นของคุณ…",
+    "thead": {
+      "amount": "จำนวน",
+      "item": "สินค้า",
+      "price": "ราคา"
+    },
+    "tbody": {
+      "shipping": "ค่าขนส่ง"
+    },
+    "tfoot": {
+      "total": "รวม"
+    }
+  }
+}
